requirements.txt
pymupdf>=1.23.0
langchain>=0.1.0
langchain-text-splitters>=0.0.1
psycopg2-binary>=2.9.9
pgvector>=0.2.5
python-dotenv>=1.0.1
sentence-transformers>=2.5.1
google-genai>=0.1.0
torchvision>=0.16.0
streamlit>=1.31.0
watchdog>=4.0.0
